+++ input-capture/src/dummy.rs
@@ -5,6 +5,7 @@ use std::time::Duration;
 
 use async_trait::async_trait;
 use futures_core::Stream;
+use input_event::CrossingModifier;
 use input_event::PointerEvent;
 use tokio::time::{self, Instant, Interval};
 
@@ -43,6 +44,14 @@ impl Capture for DummyInputCapture {
     }
 
     async fn release(&mut self, _warp_target: Option<(i32, i32)>) -> Result<(), CaptureError> {
+        Ok(())
+    }
+
+    async fn set_crossing_modifier(
+        &mut self,
+        _pos: Position,
+        _modifier: Option<CrossingModifier>,
+    ) -> Result<(), CaptureError> {
         Ok(())
     }